# Optimized the management of the assets in ebrisk by using pandas