Part of https://github.com/gem/oq-engine/issues/5454. Unfortunately there is no speedup :-(
Before:
```
   calc_48532                                  time_sec memory_mb counts     
   =========================================== ======== ========= ===========
   total ebrisk                                23_638   74        4_260      
   getting assets                              7_829    316       3_851      
   computing risk                              6_370    0.0       324_541    
   EbriskCalculator.run                        2_464    1_555     1          
   aggregating losses                          1_866    0.0       262_452_194
   getting crmodel                             1_257    1.80469   3_851      
   getting hazard                              215      0.0       11_496      
```
After:
``` 
   calc_48623                                  time_sec memory_mb counts 
   =========================================== ======== ========= =======
   total ebrisk                                23_329   64        4_260  
   aggregating losses                          8_173    0.0       324_541
   computing risk                              6_827    0.0       324_541
   getting assets                              3_539    110       3_851  
   EbriskCalculator.run                        2_428    2_146     1      
   getting crmodel                             1_263    1.45312   3_851  
   getting hazard                              228      0.0       11_496 
```